Separate C level init to own function
Instead of defining the whole class __init__ as NotImplemented, which
leads to the compiler not generating any part of the class struct, we
let the compiler generate all that code for us and instead call an
internal _init_inflater / _init_deflater method that does the actual
initialization of the _inflater / _deflater pointer. While it is a u64,
it is still boxed, so we need to use the fromB_u64 & toB_u64 to read &
set it.

Also just creating a inflater / deflater on top of these functions so we
get that nice wrapping just on the top of the function and the rest of
the function body can remain in idiomatic C style, referring to just
"inflater" in function calls etc.